Add files via upload
n this example, __name is a private attribute and _get_grade is a protected method. Private attributes and methods are denoted by a double underscore in front of the name. Protected methods and attributes are denoted by a single underscore.

We create an object of the class Student and access the public method display_student_info to display the information of the student. We can not access the private attributes or methods directly, but they can be accessed within the class or within a subclass if needed.