Add files via upload
In this example, the Shape class defines an abstract method area that must be implemented by its child classes. The Rectangle and Circle classes inherit from the Shape class and define their own implementations of the area method. The Shape class also contains a regular method display_area that prints the area of the shape.

Finally, we create objects of the Rectangle and Circle classes and call the display_area method on each of them to display the area of the shapes.